backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
asyncpg>=0.29.0
alembic>=1.13.3
pydantic>=2.10.4
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.12
boto3>=1.35.40
botocore>=1.35.40
python-dotenv>=1.0.1
email-validator>=2.0.0
celery[redis]>=5.4.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
